[aerial_robot_model][pinocchio] rename header file
