+++ aerial_robot_model/include/aerial_robot_model/model/pinocchio_robot_model.h
